Add Logic (Unfinished)
Add content in AskInfo.java. I try to separate the discovery service and the mode Ai uses when ask user the information of a given car. The action AI should do are implemented in AskInfo.java. This class will but has not connected to any services right now.

In DiscoveryGUI.java, I add a field DiscoveryGUI.returnDestination, which can help program to go back to the start when user click "Return"